Convert centibel attenuation to gain from a double, not a float
The gain lookup is indexed by whole centibels, so what matters is which side of
an integer the value falls on.  Every caller computes its centibel value as a
double and the conversion narrowed it to float first, which moved values
sitting just under an integer up onto it and picked the neighbouring table
entry.  The envelope's decay is a staircase over that table, so a step landed
one render block early and the voice carried a slightly wrong gain until the
next step caught up.

Take a double and index it directly.  Truncate toward zero rather than with
floorf, matching upstream's `| 0`; the two agree for the non-negative range the
clamp already guarantees, but the port should not have to rely on that.

Whole-note renders that were otherwise exact were losing tens of dB to this.
On the grid, note 72 goes from 62.6 dB to 99.7 dB and stops diverging at all,
the note axis as a whole from 62.6 to 87.8, attack from 72.2 to 89.3, decay
from 82.8 to 89.3 and release from 84.8 to 89.3.  Every axis outside the
repeated-note and effects pair now sits at 87 dB or better.  The corpus does
not move, mean 0.00 dB across 31 files.

+++ spessasynth_core/src/synthesizer/audio_engine/engine_components/dsp_chain/lowpass_filter.c
@@ -18,7 +18,7 @@
 #define FILTER_SMOOTHING_FACTOR 0.03f
 
 extern float ss_abs_cents_to_hz(int cents);
-extern float ss_centibel_attenuation_to_gain(float cb);
+extern float ss_centibel_attenuation_to_gain(double cb);
 
 /* ── Cached coefficient table ───────────────────────────────────────────────
  * Indexed as cache[resonance_cb][cutoff_cents_floor].
